+++ src/Glpi/System/Status/StatusChecker.php
@@ -478,8 +478,18 @@ final class StatusChecker
                 foreach ($stuck_crontasks as $ct) {
                     $status['stuck'][] = $ct['name'];
                 }
+                $errored_crontasks = array_filter($crontasks, static fn($crontask) => $crontask['state'] === CronTask::STATE_ERROR);
+                foreach ($errored_crontasks as $ct) {
+                    $status['errored'][] = $ct['name'];
+                }
                 $status['status'] = count($status['stuck']) ? self::STATUS_PROBLEM : self::STATUS_OK;
-                $status['status_msg'] = sprintf(_x('glpi_status', 'RUNNING: %d, STUCK: %d, TOTAL: %d'), $running, count($stuck_crontasks), count($crontasks));
+                $status['status_msg'] = sprintf(
+                    _x('glpi_status', 'RUNNING: %d, STUCK: %d, ERRORED: %d, TOTAL: %d'),
+                    $running,
+                    count($stuck_crontasks),
+                    count($errored_crontasks),
+                    count($crontasks)
+                );
             }
             self::$cached_status['crontasks'] = $status;
         }